move fundscard to dashboard
