+++ src/portal/views/Account.vue
@@ -115,11 +115,11 @@ import {
 } from "../lib/accepttc";
 import WelcomeWindow from "../../components/WelcomeWindow.vue";
 import { activateThroughActivationService } from "../lib/activation";
-import FundsCard from "../../components/FundsCard.vue";
+
 import Twin from "./Twin.vue";
 @Component({
   name: "AccountView",
-  components: { WelcomeWindow, FundsCard, Twin },
+  components: { WelcomeWindow, Twin },
 })
 export default class AccountView extends Vue {
   documentLink = "https://library.threefold.me/info/legal/#/";
