+++ src/portal/views/Accounts.vue
@@ -53,11 +53,10 @@ import { accountInterface } from "../store/state";
 import WelcomeWindow from "../../components/WelcomeWindow.vue";
 import Account from "./Account.vue";
 
-import FundsCard from "@/components/FundsCard.vue";
 import Twin from "./Twin.vue";
 @Component({
   name: "AccountsView",
-  components: { WelcomeWindow, Account, FundsCard, Twin },
+  components: { WelcomeWindow, Account, Twin },
 })
 export default class AccountsView extends Vue {
   searchAccount = "";
