+++ src/portal/views/Farms.vue
@@ -1,7 +1,7 @@
 <template>
   <div>Farms
     Lorem ipsum dolor sit amet consectetur, adipisicing elit. Reiciendis ut consequatur vitae enim ab magnam recusandae, quam necessitatibus amet a nulla similique perferendis officia quidem debitis aperiam dignissimos veritatis explicabo.
-    {{$store.state.portal.name}}
+
   </div>
 </template>
